+++ modules/hydra/default.nix
@@ -38,6 +38,7 @@ let
     "github:numtide/flake-utils"
     "github:numtide/nix-filter"
     "github:numtide/treefmt-nix"
+    "github:oddlama/nix-topology"
     "github:ryantm/agenix"
     "github:SenchoPens/base16.nix"
     "github:SenchoPens/fromYaml"