+++ flake.nix
@@ -313,7 +313,75 @@
 
           # Pre-commit hooks to enforce formatting, lining, find 
           # antipatterns and ensure they don't reach upstream
-          checks = import ./checks { inherit self pkgs; };
+          checks = {
+            pre-commit-hooks = self.inputs.pre-commit-hooks.lib.${system}.run {
+              src = self;
+              hooks = {
+                # Builtin hooks
+                actionlint.enable = true;
+                conform.enable = true;
+                deadnix = {
+                  enable = true;
+                  settings.edit = true;
+                };
+                nixfmt = {
+                  enable = true;
+                  settings.width = 80;
+                };
+                prettier = {
+                  enable = true;
+                  settings = {
+                    ignore-path = [ self.packages.${system}.prettierignore ];
+                    write = true;
+                  };
+                };
+
+                typos = {
+                  enable = true;
+                  settings = {
+                    binary = false;
+                    locale = "en-au";
+                  };
+                };
+
+                # Custom hooks
+                git-cliff = {
+                  enable = true;
+                  name = "Git Cliff";
+                  entry =
+                    "${pkgs.git-cliff}/bin/git-cliff --output CHANGELOG.md";
+                  language = "system";
+                  pass_filenames = false;
+                };
+
+                statix-write = {
+                  enable = true;
+                  name = "Statix Write";
+                  entry = "${pkgs.statix}/bin/statix fix";
+                  language = "system";
+                  pass_filenames = false;
+                };
+
+                trufflehog-verified = {
+                  enable = pkgs.stdenv.isLinux;
+                  name = "Trufflehog Search";
+                  entry =
+                    "${pkgs.trufflehog}/bin/trufflehog git file://. --since-commit HEAD --only-verified --fail --no-update";
+                  language = "system";
+                  pass_filenames = false;
+                };
+
+                trufflehog-regex = {
+                  enable = pkgs.stdenv.isLinux;
+                  name = "Trufflehog Regex Search";
+                  entry =
+                    "${pkgs.trufflehog}/bin/trufflehog git file://. --since-commit HEAD --config .trufflehog/config.yaml --fail --no-verification -x ./.trufflehog/path_exclusions  --no-update";
+                  language = "system";
+                  pass_filenames = false;
+                };
+              };
+            };
+          };
 
           # Shell environments (applied to both nix develop and nix-shell via
           # shell.nix in top level directory)