+++ hosts/gastly/default.nix
@@ -18,6 +18,7 @@ in {
   imports = with self.nixosModules; [
     agenix
     clamav
+    grafana-agent
     nix-topology
     fonts
     generations