+++ hosts/meowth/default.nix
@@ -1,6 +1,7 @@
 { config, self, ... }: {
   imports = with self.nixosModules; [
     agenix
+    grafana-agent
     microvm-guest
     prowlarr
     nix-topology