+++ hosts/ninetales/default.nix
@@ -45,6 +45,7 @@ in {
     fonts
     nix-topology
     generations
+    grafana-agent
     gnupg
     lorri
     networking