+++ hosts/porygon/default.nix
@@ -3,9 +3,10 @@
 
   imports = with self.nixosModules; [
     agenix
-    palworld
+    grafana-agent
     microvm-guest
     nix-topology
+    palworld
     time
     timesyncd
   ];