+++ modules/grafana-agent/default.nix
@@ -4,7 +4,11 @@ let
 
   blocky-enabled = config.services.blocky.enable;
   clamav-enabled = config.services.clamav.daemon.enable;
+  mysql-enabled = config.services.prometheus.exporters.mysqld.enable;
+  nextcloud-enabled = config.services.prometheus.exporters.redis.enable;
   node-enabled = config.services.prometheus.exporters.node.enable;
+  redis-enabled = config.services.prometheus.exporters.redis.enable;
+  # TODO: change this so it checks for zfs also.
   telegraf-enabled = config.services.telegraf.enable;
 
   blocky-prom-config = {
@@ -33,6 +37,32 @@ let
     }];
   };
 
+  mysql-prom-config = {
+    job_name = "mysql";
+    scheme = "http";
+    static_configs = [{
+      labels.host = config.networking.hostName;
+      targets = [
+        "127.0.0.1:${
+          builtins.toString config.services.prometheus.exporters.mysqld.port
+        }"
+      ];
+    }];
+  };
+
+  nextcloud-prom-config = {
+    job_name = "nextcloud";
+    scheme = "http";
+    static_configs = [{
+      labels.host = config.networking.hostName;
+      targets = [
+        "127.0.0.1:${
+          builtins.toString config.services.prometheus.exporters.nextcloud.port
+        }"
+      ];
+    }];
+  };
+
   node-prom-config = {
     job_name = "node";
     scheme = "http";
@@ -41,6 +71,19 @@ let
       targets = [
         "127.0.0.1:${
           builtins.toString config.services.prometheus.exporters.node.port
+        }"
+      ];
+    }];
+  };
+
+  redis-prom-config = {
+    job_name = "redis";
+    scheme = "http";
+    static_configs = [{
+      labels.host = config.networking.hostName;
+      targets = [
+        "127.0.0.1:${
+          builtins.toString config.services.prometheus.exporters.redis.port
         }"
       ];
     }];
@@ -105,7 +148,10 @@ in {
         configs = [{
           name = "prometheus_scrape_configs";
           scrape_configs = (lib.optional blocky-enabled blocky-prom-config)
+            ++ (lib.optional mysql-enabled mysql-prom-config)
+            ++ (lib.optional nextcloud-enabled nextcloud-prom-config)
             ++ (lib.optional node-enabled node-prom-config)
+            ++ (lib.optional redis-enabled redis-prom-config)
             ++ (lib.optional telegraf-enabled zfs-telegraf-config);
         }];