+++ hosts/wigglytuff/default.nix
@@ -26,11 +26,13 @@ in {
     agenix
     generations
     gnupg
-    nix-topology
+    grafana-agent
     greetd
     hyprland
+    i18n
     lorri
     nix
+    nix-topology
     openssh
     self.inputs.nixos-hardware.nixosModules.raspberry-pi-4
     sudo